+++ Global/startup.js
@@ -1,0 +1,17 @@
+// Startup function
+// Noah D.
+// 10-12-2025
+// 10-12-2025
+/*
+	Code to run on the startup of the program
+	Contains globally important variables such as `state`
+*/
+/// <reference path="/home/aurora/.vscode/extensions/samplavigne.p5-vscode-1.2.16/p5types/global.d.ts" />
+
+/**
+ * Global state of the game
+ * @property {number} screen Current screen ID the game is on
+ */
+var state = {
+	screen: SCREEN_IDS.TITLE_CARD,
+}