+++ Screens/win.js
@@ -1,42 +1,38 @@
 // Win function
 // Noah D.
-// 9/12/2025
-// 9/12/2025
+// 9-12-2025
+// 10-12-2025
 /*
+	ScreenID: 05
 	Function to draw the win screen
 */
 /// <reference path="/home/aurora/.vscode/extensions/samplavigne.p5-vscode-1.2.16/p5types/global.d.ts" />
 
-var bNextLevel;
-var bTryAgain;
-var bMainMenu;
-function setup() {
-	createCanvas(480, 270);
-	bNextLevel = new Button(
+var bNextLevel05;
+var bTryAgain05;
+var bMainMenu05;
+function setup05() {
+	bNextLevel05 = new Button(
 		() => { /* Go to next level */ },
 		"Next level"
 	);
-	bRetry = new Button(
+	bRetry01 = new Button(
 		() => { /* Back to beginning of level */ },
 		"Retry"
 	);
-	bMainMenu = new Button(
-		() => { /* Go to main menu */ },
+	bMainMenu05 = new Button(
+		() => { state.screen = SCREEN_IDS.MAIN_MENU; },
 		"Main menu"
 	);
 }
 
-function draw() {
-	win();
-}
-
-function win() {
+function draw05() {
 	background(220);
 	var tmp = new TextStyle(null, BOLD, 32, null, CENTER, CENTER);
 	style(tmp);
 	text("You've Won!!", width/2, height/8);
 
-	button(bNextLevel, width*3/16, height/3, width/4, height/6);
-	button(bRetry, width*9/16, height/3, width/4, height/6);
-	button(bMainMenu, width*3/8, height*2/3, width/4, height/6);
+	button(bNextLevel05, width*3/16, height/3, width/4, height/6);
+	button(bRetry01, width*9/16, height/3, width/4, height/6);
+	button(bMainMenu05, width*3/8, height*2/3, width/4, height/6);
 }